+++ src/Shared/layout/layout.js
@@ -1,18 +1,22 @@
 class LayoutHelper {
 
+  static elClassList = document.body.classList;
+
   static sidebarToggle(toggle) {
-    const minimize = arguments.length ? toggle : !document.body.classList.contains('sidebar-minimized');
+    const minimize = arguments.length ? toggle : !this.elClassList.contains('sidebar-minimized');
     this.sidebarMinimize(minimize);
     this.brandMinimize(minimize);
     this.sidebarPSToggle(!minimize);  /*remove PS on sidebar minimized*/
   }
 
   static sidebarMinimize(force) {
-    return document.body.classList.toggle('sidebar-minimized', force);
+    // return this.elClassList.toggle('sidebar-minimized', force);
+    return this.toggleClass('sidebar-minimized', force);
   }
 
   static brandMinimize(force) {
-    document.body.classList.toggle('brand-minimized', force);
+    // this.elClassList.toggle('brand-minimized', force);
+    this.toggleClass('brand-minimized', force);
   }
 
   //  sidebar perfect scrollbar
@@ -25,6 +29,18 @@ class LayoutHelper {
         sidebar.classList.remove('ps', 'ps-container', 'ps--active-y');
       }
     }
+  }
+
+  static toggleClass(className, force) {
+
+    if (force === true) {
+      this.elClassList.add(className);
+    } else if (force === false) {
+      this.elClassList.remove(className);
+    } else {
+      this.elClassList.toggle(className);
+    }
+    return this.elClassList.contains(className);
   }
 }